+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/aura_extensions/sulu-content.js
@@ -1,3 +1,12 @@
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 define(function() {
 
     'use strict';