+++ tests/Sulu/Component/Tag/Request/TagRequestHandlerTest.php
@@ -10,7 +10,6 @@
 
 namespace Sulu\Component\Tag\Request;
 
-use Sulu\Bundle\TagBundle\Entity\Tag;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
@@ -65,8 +64,7 @@ class TagRequestHandlerTest extends \PHPUnit_Framework_TestCase
      */
     public function testAppendTagToUrl($tagsParameter, $url, $tagsString, $expected)
     {
-        $tag = new Tag();
-        $tag->setName('Test');
+        $tag = ['name' => 'Test'];
 
         $requestStack = $this->prophesize(RequestStack::class);
         $request = $this->prophesize(Request::class);
@@ -100,8 +98,7 @@ class TagRequestHandlerTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetTagToUrl($tagsParameter, $url, $tagsString, $expected)
     {
-        $tag = new Tag();
-        $tag->setName('Test');
+        $tag = ['name' => 'Test'];
 
         $requestStack = $this->prophesize(RequestStack::class);
         $request = $this->prophesize(Request::class);
@@ -135,9 +132,6 @@ class TagRequestHandlerTest extends \PHPUnit_Framework_TestCase
      */
     public function testRemoveTagsFromUrl($tagsParameter, $url, $tagsString)
     {
-        $tag = new Tag();
-        $tag->setName('Test');
-
         $requestStack = $this->prophesize(RequestStack::class);
         $request = $this->prophesize(Request::class);